adderd learning rate scheduler in training
